Make html_error function public
Right now,
we're not supporting a lot of inputs
like radio buttons
and selects.
I wanted to use the error messages
and logic around that
for my other inputs.
By extracting this to a public function,
consumers can use that
without us having to define a formulator wrapper
for every single input type.