Support other fields that don't end in input
There are a handful of other fields that don't end in input. This adds
support for some of them.

This is still missing support for `radio_button` and `multiple_select`.
The former will require some additional code changes. The later may
require some additional considerations about the API. (SimpleForm
supports it as `collection`.