Fix override option for validate_regex (#52)
Problem
When supplying validate_regex: false on the input itself, it seems to be ignored

Solution
Looks like I made a logic error when checking if the option is enabled. When the option is false, it failed the condition and then moved on to check the application config, which defaults to true when not present. The change is to explicitly check for false in the options