Adds tags to releases (#71)
Closes https://github.com/thoughtbot/formulator/issues/38

What changed?
=============

We have a `bin/release` script that will set the new release version and
publish the package to hex.

We also want to create a tag and push it to GitHub so releases are
easier to track. This commit adds a simple tag to the release script and
pushes it to GitHub.